refactor(bootstrap): move searchpath into another new module
